# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AuthResponse:
    properties:
      csrf_token:
        type: string
    type: object
...
  dto.ErrorDTO:
    properties:
      message:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
...
  dto.RegisterRequest:
    properties:
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - name
    - password
    - phone_number
    type: object
...
  dto.ValidationErrorDTO:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  dto.ValidationErrorsDTO:
    properties:
      errors:
        items:
          $ref: '#/definitions/dto.ValidationErrorDTO'
        type: array
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Undefined Team
    url: https://github.com/go-park-mail-ru/2025_2_Undefined
  description: API сервер для чат-приложения в стиле Telegram. Позволяет регистрировать
    пользователей, управлять чатами и обмениваться сообщениями.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Undefined team API documentation of project Telegram
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя по номеру телефона и паролю через
        gRPC микросервис, создает сессию
      parameters:
      - description: Креденшиалы для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Вход выполнен успешно
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Ошибки валидации
          schema:
            $ref: '#/definitions/dto.ValidationErrorsDTO'
        "401":
          description: Неверные креденшиалы
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Аутентификация пользователя
      tags:
      - auth
  /logout:
    post:
      consumes:
      - application/json
      description: Аннулирует текущую сессию через gRPC микросервис и удаляет cookie
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя в системе через gRPC микросервис
        и создает сессию
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Ошибки валидации
          schema:
            $ref: '#/definitions/dto.ValidationErrorsDTO'
      summary: Регистрация пользователя
      tags:
      - auth
...
swagger: "2.0"